# Perhaps slightly faster trellis quants