The PR adds some optimizations to the GEMV implementation of the `IQ2_KT, IQ3_KT, IQ4_KT` quants.

On my Ryzen-7950X I don't notice much of a difference when running with 16 threads as the calculation is (nearly) memory bound. But when testing with fewer threads, I see quite significant gains in TG performance compared to the main branch. Here some results for LlaMA-3.1-8B-Instruct

### IQ2_KT

| model             |       size | threads |          test |      t/s (main)  |     t/s (PR)     |  Speedup |
| ----------------- | ---------: | ------: | ------------: | ---------------: | ---------------: | -------: |
| llama 8B IQ2_KT   |   2.77 GiB |       2 |         tg128 |      3.28 ± 0.00 |      4.11 ± 0.00 |   1.253  |
| llama 8B IQ2_KT   |   2.77 GiB |       4 |         tg128 |      6.28 ± 0.01 |      7.86 ± 0.00 |   1.251  |
| llama 8B IQ2_KT   |   2.77 GiB |       8 |         tg128 |     11.38 ± 0.00 |     14.02 ± 0.01 |   1.232  |

### IQ3_KT

| model            |       size | threads |          test |       t/s (main) |      t/s (PR)    |  Speedup |
| ---------------- | ---------: | ------: | ------------: | ---------------: | ---------------: | -------: |
| llama 8B IQ3_KT  |   3.58 GiB |       2 |         tg128 |      2.87 ± 0.00 |      3.92 ± 0.00 |   1.366  |
| llama 8B IQ3_KT  |   3.58 GiB |       4 |         tg128 |      5.58 ± 0.00 |      7.50 ± 0.00 |   1.344  |
| llama 8B IQ3_KT  |   3.58 GiB |       8 |         tg128 |     10.20 ± 0.00 |     13.42 ± 0.01 |   1.316  |

### IQ4_KT

| model            |       size | threads |          test |     t/s (main)   |   t/s (PR)       |  Speedup |
| ---------------- | ---------: | ------: | ------------: | ---------------: | ---------------: | -------: |
| llama 8B IQ4_KT  |   4.30 GiB |       2 |         tg128 |      2.26 ± 0.00 |      3.27 ± 0.00 |   1.447  |
| llama 8B IQ4_KT  |   4.30 GiB |       4 |         tg128 |      4.38 ± 0.00 |      6.25 ± 0.01 |   1.427  |
| llama 8B IQ4_KT  |   4.30 GiB |       8 |         tg128 |      8.11 ± 0.00 |     11.30 ± 0.00 |   1.393  |

@ubergarm 

In your performance testing on the 6980P system `iqX_kt` quants were very far from saturating memory bandwidth, so perhaps you will see bigger gains there than I see on my system when using all cores.