@ubergarm 2025-06-19
My usual library spot was closed today so sitting outside in the sun trying to grab some quick llama-sweep-bench numbers:

* 7965WX running R1-0528-IQ3_KT
  - PR541@93209939
    *  10.58 TG tok/sec
  - main@144ee1c4
    * 8.61 TG tok/sec

So on the AMD Thread Ripper Pro seeing improvementn from 8.61 up to 10.58 TG tok/sec so 1.229x improvement speedup! Great considering this is also using CUDA offload.

<details>

<summary>llama-sweep-bench command</summary>

```bash
./build/bin/llama-sweep-bench \
    --model "$model" \
    --no-mmap \
    --ctx-size 8704 \
    -ctk f16 \
    -mla 3 -fa \
    -fmoe \
    -amb 512 \
    -ngl 99 \
    -ot "blk\.(3|4|5|6|7|8|9)\.ffn_.*=CUDA0" \
    -ot "blk\.(10|11|12|13|14|15|16)\.ffn_.*=CUDA1" \
    -ot exps=CPU \
    --warmup-batch \
    --threads 24
```
main: n_kv_max = 8704, n_batch = 2048, n_ubatch = 512, flash_attn = 1, n_gpu_layers = 99, n_threads = 24, n_threads_batch = 24

## version: 3764 (93209939) (PR541)
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    6.297 |    81.30 |   12.097 |    10.58 |
|   512 |    128 |    512 |    6.421 |    79.74 |   12.570 |    10.18 |
|   512 |    128 |   1024 |    6.515 |    78.59 |   12.184 |    10.51 |
|   512 |    128 |   1536 |    7.365 |    69.52 |   12.578 |    10.18 |

## version: 3761 (144ee1c4) (main)
main: n_kv_max = 8704, n_batch = 2048, n_ubatch = 512, flash_attn = 1, n_gpu_layers = 99, n_threads = 24, n_threads_batch = 24

|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    9.495 |    53.92 |   14.858 |     8.61 |
|   512 |    128 |    512 |    9.533 |    53.71 |   14.918 |     8.58 |
|   512 |    128 |   1024 |    9.535 |    53.70 |   15.376 |     8.32 |
|   512 |    128 |   1536 |    9.911 |    51.66 |   14.961 |     8.56 |
|   512 |    128 |   2048 |   10.121 |    50.59 |   14.990 |     8.54 |
|   512 |    128 |   2560 |   10.010 |    51.15 |   15.133 |     8.46 |
|   512 |    128 |   3072 |   10.075 |    50.82 |   15.551 |     8.23 |
|   512 |    128 |   3584 |   10.190 |    50.24 |   15.575 |     8.22 |
|   512 |    128 |   4096 |   10.712 |    47.80 |   15.185 |     8.43 |
|   512 |    128 |   4608 |   10.329 |    49.57 |   15.294 |     8.37 |

</details>

I'll try to get some numbers on the big 6980P pure-CPU soon!