@ubergarm 2025-06-20
Okay, back at a desk with my laptop for a little while. Here is a quick comparison for a mixed R1-0528-IQ3_KT quant.

* Intel Xeon 6980P
* Single Socket
* CPU-Only compiled
* First line of llama-sweep-bench PP512/TG128/N_KV0
* 272.527 GiB (3.483 BPW)
- type  f32:  361 tensors
- type q5_0:   61 tensors `attn_k_b`
- type q8_0:    1 tensors `token_embd`
- type iq5_ks:  550 tensors `shexp/dense/attn`
- type iq3_kt:  116 tensors `ffn_(gate|up)_exps`
- type iq4_kt:   58 tensors `ffn_down_exps`

| TG tok/sec `main@144ee1c4` | TG tok/sec `PR541@93209939` | speed-up |
| --- | --- | --- |
| 6.29 | 8.22 | 1.309x |

Given not every tensor is `kt` type, actual speed-ups are likely higher. I don't have a good set of pure `kt`'s to test easily like you did above, but my limited testing suggests a big improvement in TG for all three `kt` quant types in both MoE and dense models.

I spot checked using less threads for TG and it was slower, so using `--threads 128 --threads-batch 128` seemed best. There is some little variation in multiple runs as well.

Finally, I didn't expect this, but it seems like PP increased *a lot* as well!!?? At the default batch size PP went from 36.75 up to 117.38,  a ~3.19x speedup!!? I didn't track the code-path to see if the new avx512 and other code is used for PP as well as TG? The effect is not as dramatic at higher batch sizes, but still holds as being faster at ub 4096.

No graphs, tonight, but some data in the fold below showing the effect.

<details>

<summary>👈 llama-sweep-bench command and data </summary>

```bash
model=/mnt/raid/models/ubergarm/DeepSeek-R1-0528-GGUF/IQ3_KT/DeepSeek-R1-0528-IQ3_KT-00001-of-00006.gguf

# adjust -ub 2048 -b 2048
# also adjust -c to large enough for batch size or it will segfault out

numactl -N 0 -m 0 \
./build/bin/llama-sweep-bench \
    --model "$model" \
    -c 1536 \
    -ctk q8_0 \
    -mla 3 -fa \
    -fmoe \
    --no-mmap \
    --threads 128 \
    --threads-batch 128 \
    --numa numactl \
    --warmup-batch
```

## main@144ee1c4

main: n_kv_max = 1536, n_batch = 2048, n_ubatch = 512, flash_attn = 1, n_gpu_layers = -1, n_threads = 128, n_threads_batch = 128
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |   13.957 |    36.68 |   20.376 |     6.28 |
|   512 |    128 |    512 |   14.501 |    35.31 |   20.703 |     6.18 |
|   512 |    128 |   1024 |   14.865 |    34.44 |   22.955 |     5.58 |

main: n_kv_max = 6144, n_batch = 2048, n_ubatch = 2048, flash_attn = 1, n_gpu_layers = -1, n_threads = 128, n_threads_batch = 128
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  2048 |    512 |      0 |   24.211 |    84.59 |   83.107 |     6.16 |
|  2048 |    512 |   2048 |   27.291 |    75.04 |   94.896 |     5.40 |
|  2048 |    512 |   4096 |   30.650 |    66.82 |   95.730 |     5.35 |

main: n_kv_max = 12288, n_batch = 4096, n_ubatch = 4096, flash_attn = 1, n_gpu_layers = -1, n_threads = 128, n_threads_batch = 128

|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  4096 |   1024 |      0 |   37.711 |   108.62 |  174.245 |     5.88 |
|  4096 |   1024 |   4096 |   49.629 |    82.53 |  196.952 |     5.20 |
|  4096 |   1024 |   8192 |   59.777 |    68.52 |  199.099 |     5.14 |

---

## PR541@93209939 

main: n_kv_max = 1536, n_batch = 2048, n_ubatch = 512, flash_attn = 1, n_gpu_layers = -1, n_threads = 128, n_threads_batch = 128
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    4.362 |   117.38 |   15.564 |     8.22 |
|   512 |    128 |    512 |    4.729 |   108.26 |   17.158 |     7.46 |
|   512 |    128 |   1024 |    4.942 |   103.60 |   19.407 |     6.60 |

main: n_kv_max = 6144, n_batch = 2048, n_ubatch = 2048, flash_attn = 1, n_gpu_layers = -1, n_threads = 128, n_threads_batch = 128
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  2048 |    512 |      0 |   14.727 |   139.07 |   65.669 |     7.80 |
|  2048 |    512 |   2048 |   18.297 |   111.93 |   81.433 |     6.29 |
|  2048 |    512 |   4096 |   21.476 |    95.36 |   82.792 |     6.18 |

main: n_kv_max = 12288, n_batch = 4096, n_ubatch = 4096, flash_attn = 1, n_gpu_layers = -1, n_threads = 128, n_threads_batch = 128

|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|  4096 |   1024 |      0 |   31.189 |   131.33 |  137.583 |     7.44 |
|  4096 |   1024 |   4096 |   42.929 |    95.41 |  162.857 |     6.29 |
|  4096 |   1024 |   8192 |   53.700 |    76.28 |  149.823 |     6.83 |

</details>

fwiw here is the output of `lscpu | grep Flags` on the 6980P

<details>

<summary>👈 6980P CPU flags</summary>

```
$ lscpu | grep Flags
Flags:                                fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf tsc_known_freq pni pclmulqdq dtes64 monitor ds_cpl vmx smx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid dca sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb cat_l3 cat_l2 cdp_l3 intel_ppin cdp_l2 ssbd mba ibrs ibpb stibp ibrs_enhanced tpr_shadow flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm cqm rdt_a avx512f avx512dq rdseed adx smap avx512ifma clflushopt clwb intel_pt avx512cd sha_ni avx512bw avx512vl xsaveopt xsavec xgetbv1 xsaves cqm_llc cqm_occup_llc cqm_mbm_total cqm_mbm_local split_lock_detect user_shstk avx_vnni avx512_bf16 wbnoinvd dtherm ida arat pln pts hwp hwp_act_window hwp_epp hwp_pkg_req vnmi avx512vbmi umip pku ospke waitpkg avx512_vbmi2 gfni vaes vpclmulqdq avx512_vnni avx512_bitalg avx512_vpopcntdq la57 rdpid bus_lock_detect cldemote movdiri movdir64b enqcmd fsrm md_clear serialize tsxldtrk pconfig arch_lbr ibt amx_bf16 avx512_fp16 amx_tile amx_int8 flush_l1d arch_capabilities
```

</details>

Thanks!